requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
cryptography==41.0.4
Django==4.2.4
django-password-validators==1.7.1
django-paypal==2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
idna==3.4
Pillow==10.1.0 
psycopg2==2.9.7
pycparser==2.21
PyJWT==1.7.1
pyotp==2.9.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
types-requests==2.31.0.10
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4